add floyd

+++ Contents/3_Graph/1_Dijkstra.cpp
@@ -1,16 +1,16 @@
 // 0-base
-const long long INF = 1e18;
+const LL INF = 1e18;
 const int MAXN = ;
 struct Edge {
   int at;
-  long long cost;
+  LL cost;
   bool operator < (const Edge &other) const {
     return cost > other.cost;
   }
 };
 
 int n;
-long long dis[MAXN];
+LL dis[MAXN];
 vector<Edge> G[MAXN];
 
 void init() {
